Move PureInterpreterState into a separate file
