+++ libyul/tools/interpreter/PureInterpreterState.h
@@ -1,0 +1,41 @@
+/*
+	This file is part of solidity.
+
+	solidity is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version.
+
+	solidity is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with solidity.  If not, see <http://www.gnu.org/licenses/>.
+*/
+// SPDX-License-Identifier: GPL-3.0
+
+#pragma once
+
+#include <cstddef>
+
+namespace solidity::yul::tools::interpreter
+{
+
+struct PureInterpreterConfig
+{
+	size_t maxTraceSize = 0;
+	size_t maxSteps = 0;
+	size_t maxExprNesting = 0;
+	size_t maxRecursionDepth = 0;
+};
+
+struct PureInterpreterState
+{
+	PureInterpreterConfig const config;
+
+	size_t numSteps = 0;
+};
+
+}
